+++ R/is_gitlab.R
@@ -1,0 +1,1 @@
+is_gitlab = function() nchar(Sys.getenv("GITLAB_CI")) != 0